backend/requirements/base.txt
Django==5.1.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-ckeditor==6.7.3
django-js-asset==3.1.2
gunicorn==23.0.0
psycopg[binary]==3.2.6
Pillow==11.2.1
PyJWT==2.10.1
django-redis==5.4.0
django-ratelimit==4.1.0
